backend/requirements.txt
fastapi>=0.115.6,<0.116
uvicorn[standard]>=0.32.0,<0.33
python-multipart>=0.0.9
sqlalchemy>=2.0.36,<2.1
alembic>=1.14.0,<2
psycopg[binary]>=3.2.0,<4
openai>=1.57.0,<2
anthropic>=0.45.0,<1
httpx>=0.27.0,<1
requests>=2.32.0,<3
pydantic>=2.11.0,<3
pydantic-settings>=2.6.0,<3
google-auth>=2.36.0,<3
PyJWT>=2.9.0,<3
